IntervalTree optimizations + RDD instrumentation (#5)
* groupbykey to reducebykey

* Added RDD instrumentation

+++ src/main/scala/genApp/Interval.scala
@@ -1,6 +1,6 @@
 package genApp
 
-case class Interval[T <% Long](start: T, end: T) {
+case class Interval[T <% Int](start: T, end: T) {
   def overlaps(other: Interval[T]): Boolean = {
     (end >= start) && (other.end >= other.start) &&
       (end >= other.start && start <= other.end)
