+++ src/main/scala/genApp/IntervalTreeJoin.scala
@@ -46,19 +46,21 @@ case class IntervalTreeJoin(left: SparkPlan,
     val v1kv = v1.map(x => {
       val v1Key = buildKeyGenerator(x)
 
-      (new Interval[Long](v1Key.getLong(0), v1Key.getLong(1)),
+      (new Interval[Int](v1Key.getInt(0), v1Key.getInt(1)),
         x.copy())
     })
     val v2 = right.execute()
     val v2kv = v2.map(x => {
       val v2Key = streamKeyGenerator(x)
-      (new Interval[Long](v2Key.getLong(0), v2Key.getLong(1)),
+      (new Interval[Int](v2Key.getInt(0), v2Key.getInt(1)),
         x.copy())
     })
     /* As we are going to collect v1 and build an interval tree on its intervals,
     make sure that its size is the smaller one. */
     if (v1.count <= v2.count) {
-      val v3 = IntervalTreeJoinImpl.overlapJoin(context.sparkContext, v1kv, v2kv).flatMap(l => l._2.map(r => (l._1, r)))
+      val v3 = IntervalTreeJoinImpl.overlapJoin(context.sparkContext, v1kv, v2kv)
+        .flatMap(l => l._2
+        .map(r => (l._1, r)))
       v3.map {
         case (l: InternalRow, r: InternalRow) => {
           val joiner = GenerateUnsafeRowJoiner.create(left.schema, right.schema);
