+++ src/main/scala/ncl/Backpack.scala
@@ -1,5 +1,5 @@
 package ncl
 
-case class Backpack[T](intervalList: List[(Interval[Long],T)], processedInterval: Interval[Long]) {
+case class Backpack[T](intervalList: List[(Interval[Int],T)], processedInterval: Interval[Int]) {
 
 }
