+++ src/main/scala/ncl/Interval.scala
@@ -1,5 +1,5 @@
 package ncl
 
-case class Interval[T <% Long](start: T, end: T) {
+case class Interval[T <% Int](start: T, end: T) {
 
 }
