+++ src/main/scala/ncl/NCListBuilder.scala
@@ -3,13 +3,15 @@ package ncl
 import scala.collection.mutable
 
 object NCListBuilder {
-  def build[T](list: List[(Interval[Long], T)]): NCList = {
+  def build[T](list: List[(Interval[Int], T)]): NCList = {
 
     var topNCList = NCList(Array.empty[NCList], 0, Array.empty[Int])
     var landingNCList = NCList(Array.empty[NCList], 0, Array.empty[Int])
 
     var listWithIndices = list.zipWithIndex.map{case (k,v) => (v,k)}
-    var sortedIndices = listWithIndices.sortWith((x, y) => x._2._1.end > y._2._1.end).sortWith((x, y) => x._2._1.start < y._2._1.start).map(x => x._1)
+    var sortedIndices = listWithIndices.sortWith((x, y) => x._2._1.end > y._2._1.end)
+      .sortWith((x, y) => x._2._1.start < y._2._1.start)
+      .map(x => x._1)
 
     var stack = mutable.Stack[NCListBuildingStack]()
